==> analysis/07D_Combined_Figures.Rmd <==
---
title: "Combined Figures"
output:
    workflowr::wflow_html:
        code_folding: hide
        df_print: paged
---

```{r knitr, include = FALSE}
DOCNAME = "07D_Combined_Figures"
knitr::opts_chunk$set(autodep        = TRUE,
                      cache          = TRUE,
                      cache.path     = paste0("cache/", DOCNAME, "/"),
                      cache.comments = FALSE,
                      echo           = TRUE,
                      error          = FALSE,
                      fig.align      = "center",
                      fig.width      = 10,
                      fig.height     = 8,
                      message        = FALSE,
                      warning        = FALSE)
```

```{r libaries, cache = FALSE}
# scRNA-seq
library("Seurat")
library("monocle")

# Plotting
library("clustree")
library("cowplot")

# Presentation
library("glue")
library("knitr")

# Parallel

# Paths
library("here")

# Output

# Tidyverse
library("tidyverse")
```

```{r source, cache = FALSE}
source(here("R/output.R"))
source(here("R/crossover.R"))
```

```{r paths}
comb.path <- here("data/processed/Combined_clustered.Rds")
comb.neph.path <- here("data/processed/Combined_nephron.Rds")

o.path <- here("output/04_Organoids_Clustering/cluster_assignments.csv")
on.path <- here("output/04B_Organoids_Nephron/cluster_assignments.csv")
c.path <- here("output/07_Combined_Clustering/cluster_assignments.csv")
cn.path <- here("output/07B_Combined_Nephron/cluster_assignments.csv")

dir.create(here("output", DOCNAME), showWarnings = FALSE)
```

Introduction
============

In this document we are going to look at all of the organoids analysis results
and produce a series of figures for the paper.

```{r load-comb, cache.extra = tools::md5sum(comb.path), cache.lazy = FALSE}
if (file.exists(comb.path)) {
    comb <- read_rds(comb.path)
} else {
    stop("Clustered Combined dataset is missing. ",
         "Please run '07_Combined_Clustering.Rmd' first.",
         call. = FALSE)
}
```

```{r load-comb-neph, cache.extra = tools::md5sum(comb.neph.path)}
if (file.exists(comb.neph.path)) {
    comb.neph <- read_rds(comb.neph.path)
} else {
    stop("Clustered Combined nephron dataset is missing. ",
         "Please run '07B_Organoids_Nephron.Rmd' first.",
         call. = FALSE)
}
```

```{r load-clusts, cache.extra = c(tools::md5sum(o.path), tools::md5sum(on.path), tools::md5sum(c.path), tools::md5sum(cn.path))}
orgs.clusts <- read_csv(o.path,
                        col_types = cols(
                            Cell = col_character(),
                            Dataset = col_character(),
                            Sample = col_integer(),
                            Barcode = col_character(),
                            Cluster = col_integer()
                        )) %>%
    rename(Organoids = Cluster)

orgs.neph.clusts <- read_csv(on.path,
                             col_types = cols(
                                 Cell = col_character(),
                                 Dataset = col_character(),
                                 Sample = col_integer(),
                                 Barcode = col_character(),
                                 Cluster = col_integer()
                             )) %>%
    rename(OrgsNephron = Cluster)

comb.clusts <- read_csv(c.path,
                        col_types = cols(
                            Cell = col_character(),
                            Dataset = col_character(),
                            Sample = col_integer(),
                            Barcode = col_character(),
                            Cluster = col_integer()
                        )) %>%
    rename(Combined = Cluster)

comb.neph.clusts <- read_csv(cn.path,
                             col_types = cols(
                                 Cell = col_character(),
                                 Dataset = col_character(),
                                 Sample = col_integer(),
                                 Barcode = col_character(),
                                 Cluster = col_integer()
                             )) %>%
    rename(CombNephron = Cluster)

clusts <- comb.clusts %>%
    left_join(comb.neph.clusts,
              by = c("Cell", "Dataset", "Sample", "Barcode")) %>%
    left_join(orgs.clusts,
              by = c("Cell", "Dataset", "Sample", "Barcode")) %>%
    left_join(orgs.neph.clusts,
              by = c("Cell", "Dataset", "Sample", "Barcode"))
```

Figure 2
========

Figure 2A
---------

```{r fig-2A}
plot.data <- comb %>%
    GetDimReduction("tsne",  slot = "cell.embeddings") %>%
    data.frame() %>%
    rownames_to_column("Cell") %>%
    mutate(Group = comb@meta.data$Group) %>%
    mutate(Group = if_else(Group == "Lindstrom", "Fetal kidney", Group)) %>%
    mutate(Cluster = comb@ident)

f2A <- ggplot(plot.data, aes(x = tSNE_1, y = tSNE_2, colour = Group)) +
    geom_point() +
    scale_color_manual(values = c("#EC008C", "#00ADEF")) +
    guides(colour = guide_legend(ncol = 1,
                                 override.aes = list(size = 10))) +
    theme_cowplot() +
    theme(legend.position = "bottom",
          legend.title = element_blank(),
          legend.justification = "center")

ggsave(here("output", DOCNAME, "figure2A.png"), f2A,
       height = 8, width = 10)
ggsave(here("output", DOCNAME, "figure2A.pdf"), f2A,
       height = 8, width = 10)

f2A
```

Figure 2B
---------

```{r fig-2B}
lab.data <- plot.data %>%
    group_by(Cluster) %>%
    summarise(tSNE_1 = mean(tSNE_1),
              tSNE_2 = mean(tSNE_2)) %>%
    mutate(Label = paste0("C", Cluster))

clust.labs <- c(
    "C0 (Stroma)\nTAGLN, ALDH1A2, COL6A3",
    "C1 (Stroma)\nDCN, LUM, POSTN",
    "C2 (Stroma)\nMGP, GATA3, NDUFA4L2",
    "C3 (Stroma)\nZFHX4, CTSK, HNRNPA1",
    "C4 (Endothelium)\nGNG11, CLDN5, CD34",
    "C5 (Cell cycle)\nHIST1H4C, PCLAF, DUT",
    "C6 (Committing nephron prog.)\nLYPD1, DAPL1, PAX8",
    "C7 (Podocyte)\nPODXL, MAFB, TCF21",
    "C8 (Cell cycle)\nCENPF, HMGB2, UBE2C",
    "C9 (Stroma)\nCOL9A3, FIBIN, DLK1",
    "C10 (Nephron)\nIGFBP7, EMBX2, MAL",
    "C11 (Glial)\nTTYH1, FABP7, SOX2",
    "C12 (Immune)\nPTPRC, TYROBP, FCER1G",
    "C13 (Blood)\nHBG1, HBB, HBA2",
    "C14 (Neural prog.)\nSOX11, ELAVL3, ELAVL4",
    "C15 (Podocyte)\nPTPRO, TCF21, PODXL"
)

f2B <- ggplot(plot.data, aes(x = tSNE_1, y = tSNE_2, colour = Cluster)) +
    geom_point() +
    geom_text(data = lab.data, aes(label = Label), colour = "black", size = 8) +
    scale_colour_discrete(labels = clust.labs) +
    guides(colour = guide_legend(nrow = 3, override.aes = list(size = 10),
                                 label.theme = element_text(size = 10))) +
    theme_cowplot() +
    theme(legend.position = "bottom",
          legend.title = element_blank(),
          legend.justification = "center")

ggsave(here("output", DOCNAME, "figure2B.png"), f2B,
       height = 8, width = 10)
ggsave(here("output", DOCNAME, "figure2B.pdf"), f2B,
       height = 8, width = 10)

f2B
```

Figure 2AB legend
-----------------

```{r fig-2AB-legend}
l2A <- get_legend(f2A)
l2B <- get_legend(f2B)

l2AB <- plot_grid(l2A, l2B, nrow = 1, rel_widths = c(0.15, 1))

l2AB
```

Figure 2C
---------

```{r fig-2C}
f2C <- summariseClusts(clusts, Combined, Organoids) %>%
    replace_na(list(Jaccard = 0)) %>%
    ggplot(aes(x = Combined, y = Organoids, fill = Jaccard)) +
    geom_tile() +
    scale_fill_viridis_c(limits = c(0, 1), name = "Jaccard\nindex") +
    scale_x_discrete(labels = paste0("C", 0:15)) +
    scale_y_discrete(labels = paste0("O", 0:12)) +
    coord_equal() +
    xlab("Combined clusters") +
    ylab("Organoid clusters") +
    theme_minimal() +
    theme(axis.text = element_text(size = 10, colour = "black"),
          axis.ticks = element_blank(),
          axis.title = element_text(size = 15),
          legend.key.height = unit(30, "pt"),
          legend.title = element_text(size = 15),
          legend.text = element_text(size = 10),
          panel.grid = element_blank())

ggsave(here("output", DOCNAME, "figure2C.png"), f2C,
       height = 8, width = 10)
ggsave(here("output", DOCNAME, "figure2C.pdf"), f2C,
       height = 8, width = 10)

f2C
```

Figure 2D
---------

```{r fig-2D}
labs <- paste0("C", 0:max(as.numeric(comb@meta.data$Cluster)))

plot.data <- comb@meta.data %>%
    select(Cluster, Group) %>%
    mutate(Cluster = paste0("C", Cluster)) %>%
    mutate(Cluster = factor(Cluster, levels = labs)) %>%
    mutate(Group = if_else(Group == "Lindstrom", "Fetal kidney", Group))

f2D <- ggplot(plot.data, aes(x = Cluster, fill = Cluster)) +
    geom_bar() +
    labs(y = "Number of cells") +
    facet_wrap(~ Group, ncol = 1, scales = "free_y") +
    theme_cowplot() +
    theme(legend.position = "none",
          axis.title.x = element_blank(),
          axis.text = element_text(size = 10),
          strip.text = element_text(size = 15,
                                    margin = margin(0, 0, 2, 0, "pt")),
          strip.background = element_blank(),
          strip.placement = "outside")

ggsave(here("output", DOCNAME, "figure2D.png"), f2D,
       height = 8, width = 10)
ggsave(here("output", DOCNAME, "figure2D.pdf"), f2D,
       height = 8, width = 10)

f2D
```

Figure 2E
---------

```{r fig-2E}
plot.data <- comb@meta.data %>%
    select(Cluster, Group) %>%
    mutate(Tissue = fct_collapse(Cluster,
                                 Stroma = c("0", "1", "2", "3", "9"),
                                 Nephron = c("6", "7", "10", "15"),
                                 Endothelial = c("4"),
                                 `Cell cycle` = c("5", "8"),
                                 Other = c("11", "12", "13", "14"))) %>%
    mutate(Tissue = fct_relevel(Tissue,
                                "Stroma", "Nephron", "Endothelial",
                                "Cell cycle", "Other")) %>%
    group_by(Group, Tissue) %>%
    summarise(Count = n()) %>%
    mutate(Prop = Count / sum(Count)) %>%
    ungroup() %>%
    mutate(Group = if_else(Group == "Lindstrom", "Fetal kidney", Group))

f2E <- ggplot(plot.data, aes(x = Group, y = Prop, fill = Tissue)) +
    geom_col() +
    labs(y = "Proportion of cells") +
    guides(fill = guide_legend(direction = "vertical", ncol = 2)) +
    theme_cowplot() +
    theme(axis.title.x = element_blank(),
          legend.title = element_blank(),
          legend.position = "bottom")

ggsave(here("output", DOCNAME, "figure2E.png"), f2E,
       height = 8, width = 10)
ggsave(here("output", DOCNAME, "figure2E.pdf"), f2E,
       height = 8, width = 10)

f2E
```

Figure 2F
---------

```{r fig-2F}
genes <- c("HSPA1A", "HSPA1B", "HBG2", "HSP90AA1", "DNAJB1", "FOS", "HSPA6",
           "HBG1", "JUN", "GPC3", "MARCKSL1", "MARCKS", "RPS26", "YBX1", 
           "HMGN2", "EEF1A1", "RPL41", "RPL39", "H3F3A", "RPS17")

gene.groups <- c(rep("Up-regulated in fetal kidney", 10),
                 rep("Up-regulated in organoid", 10)) %>%
    fct_relevel("Up-regulated in fetal kidney", "Up-regulated in organoid")
names(gene.groups) <- genes

group.labs <- c("Fetal kidney", "Organoid")

plot.data <- data.frame(FetchData(comb.neph, vars.all = genes)) %>%
    rownames_to_column("Cell") %>%
    mutate(Group = comb.neph@meta.data$Group) %>%
    gather(key = "Gene", value = "Expr", -Cell, -Group) %>%
    group_by(Group, Gene) %>%
    summarize(AvgExpr = mean(expm1(Expr)),
              PctExpr = Seurat:::PercentAbove(Expr, threshold = 0) * 100) %>%
    group_by(Gene) %>%
    mutate(AvgExprScale = scale(AvgExpr)) %>%
    mutate(AvgExprScale = Seurat::MinMax(AvgExprScale,
                                         max = 2.5, min = -2.5)) %>%
    ungroup() %>%
    mutate(GeneGroup = gene.groups[Gene]) %>%
    mutate(Gene = factor(Gene, levels = genes))

f2F <- ggplot(plot.data,
              aes(x = Gene, y = Group, size = PctExpr,
                  colour = Group, alpha = AvgExprScale)) +
       geom_point() +
       scale_radius(range = c(0, 10)) +
       scale_color_manual(values = c("#EC008C", "#00ADEF")) +
       scale_y_discrete(labels = group.labs) +
       facet_grid(~ GeneGroup, scales = "free_x") +
       theme(axis.title.x = element_blank(),
             axis.title.y = element_blank(),
             axis.text.x = element_text(angle = 90, vjust = 0.5, hjust = 1),
             panel.spacing = unit(x = 1, units = "lines"),
             strip.background = element_blank(),
             strip.placement = "outside",
             strip.text = element_text(size = 15,
                                       margin = margin(0, 0, 2, 0, "pt")),
             legend.position = "none")

ggsave(here("output", DOCNAME, "figure2F.png"), f2F,
       height = 8, width = 10)
ggsave(here("output", DOCNAME, "figure2F.pdf"), f2F,
       height = 8, width = 10)

f2F
```

Figure 2G
---------

```{r fig-2G}
genes.list <- list(
    `0` = c("GPC3", "SULT1E1", "CCBE1", "LBH", "CKS2", "STMN2", "NRK", "ACTA2",
            "PDGFRA", "ALX1", "COL6A3", "ALDH1A2", "TAGLN"),
    `1` = c("DKK2", "WNT2B", "S100A10", "SERPINI1", "PALLD", "DKK1", "ACTA2_1",
            "SFRP2", "POSTN", "IGF1", "LUM", "COL1A1", "DCN"),
    `2` = c("ANGPT2", "MCAM", "TUBA1B", "ENG", "ID1", "REN", "FTL", "SSBP4",
            "PDLIM1", "MEF2C", "PDGFRB", "GATA3", "MGP"),
    `3` = c("ROBO2", "ANGPT1", "NR2F1", "ACTA2_2", "SNHG8", "DKK1_1", "VAMP2",
            "VIM", "TBX3", "SOX4", "HNRNPA1", "CTSK", "ZFHX4"),
    `9` = c("IGF1_1", "ITM2B", "PLPP3", "COL1A1_1", "COL11A1", "COL9A1",
            "PEG10", "CNMD", "OGN", "FIBIN", "LUM_1", "DCN_1", "COL9A3"),
    `4` = c("CXCL8", "CXCL1", "CXCL2", "ICAM1", "HEY1", "CXCR4", "APLN",
            "PECAM1", "PLXND1", "CD34", "ESAM", "CLDN5", "GNG11"),
    `11` = c("PMP22", "MPZ", "ERBB3", "ARHGAP15", "MSX1", "ZFP36L1", "FABP7",
             "AP1S2", "VIM_1", "SOX2", "PLP1", "TTYH1", "S100B"),
    `14` = c("STMN1", "NCAM1", "RBP1", "ATOH1", "ELAVL4", "SOX11", "ELAVL3",
             "DCX", "TUBA1A", "STMN2_1", "CRABP1", "TUBB2B", "HES6"),
    `12` = c("CXCL8_1", "IL1B", "HLA-DPA1", "HLA-DPB1", "LGALS1", "CTSG", "MPO",
             "PRTN3", "TYROBP", "S100A9", "SRGN", "CD74", "PTPRC"),
    `13` = c("HBA1", "HBA2", "HBB", "HBG1", "AHSP", "HBM", "GNG5", "EGFL7",
             "FSCN1", "FKBP1A", "GPX1", "SLC25A3", "GYPC")
)

clust.labs <- c("C0\nStr", "C1\nStr", "C2\nStr", "C3\nStr", "C9\nStr",
                "C4\nEndo", "C11\nGlia", "C14\nNeural", "C12\nImmune",
                "C13\nBlood")

plot.data <- lapply(names(genes.list), function(name) {
    genes.raw <- genes.list[[name]]
    genes <- str_remove(genes.raw, "_[0-9]")
    names(genes.raw) <- genes
    comb %>%
        FetchData(vars.all = genes) %>%
        as_data_frame() %>%
        rownames_to_column("Cell") %>%
        mutate(Cluster = as.numeric(as.character(comb@ident)),
               Group = comb@meta.data$Group) %>%
        gather(key = "Gene", value = "Expr", -Cell, -Cluster, -Group) %>%
        group_by(Cluster, Gene, Group) %>%
        summarize(AvgExpr = mean(expm1(Expr)),
                  PctExpr = Seurat:::PercentAbove(Expr, threshold = 0) * 100) %>%
        group_by(Gene) %>%
        mutate(AvgExprScale = scale(AvgExpr)) %>%
        mutate(AvgExprScale = Seurat::MinMax(AvgExprScale,
                                             max = 2.5, min = -2.5)) %>%
        filter(Cluster == as.numeric(name)) %>%
        ungroup() %>%
        mutate(Gene = genes.raw[Gene])
})

plot.data <- plot.data %>%
    bind_rows() %>%
    mutate(Cluster = factor(Cluster,
                            levels = names(genes.list),
                            labels = clust.labs)) %>%
    mutate(Gene = factor(Gene, levels = unlist(genes.list)))

f2G <- ggplot(plot.data,
              aes(x = Group, y = Gene, size = PctExpr,
                  colour = Group, alpha = AvgExprScale)) +
    geom_point() +
    scale_radius(range = c(0, 8)) +
    scale_alpha(range = c(0.1, 1)) +
    scale_colour_manual(values = c("#EC008C", "#00ADEF")) +
    scale_x_discrete(labels = c("Fetal kidney", "Organoid")) +
    scale_y_discrete(labels = str_remove(unlist(genes.list), "_[0-9]"),
                     breaks = unlist(genes.list)) +
    facet_wrap(~ Cluster, scales = "free", nrow = 1) +
    theme(axis.title.x = element_blank(),
          axis.title.y = element_blank(),
          axis.text.x = element_text(angle = 90, vjust = 0.5, hjust = 1),
          panel.spacing = unit(x = 1, units = "lines"),
          strip.background = element_blank(),
          strip.placement = "outside",
          legend.position = "none")

ggsave(here("output", DOCNAME, "figure2G.png"), f2G,
       height = 8, width = 10)
ggsave(here("output", DOCNAME, "figure2G.pdf"), f2G,
       height = 8, width = 10)

f2G
```

Figure 2 Panel
--------------

```{r fig2-panel, fig.height = 20, fig.width = 16}
p1 <- plot_grid(f2A + theme(legend.position = "none"),
                f2B + theme(legend.position = "none"),
                nrow = 1, labels = c("A", "B"),
                label_size = 20)
p2 <- plot_grid(f2C, f2D, f2E,
                nrow = 1, rel_widths = c(1.1, 1, 0.5),
                labels = c("C", "D", "E"),
                label_size = 20)
panel <- plot_grid(p1, l2AB, p2, f2F, f2G, ncol = 1,
                   labels = c("", "", "", "F", "G"),
                   rel_heights = c(1, 0.22, 0.7, 0.4, 0.9),
                   label_size = 20)

ggsave(here("output", DOCNAME, "figure2_panel.png"), panel,
       height = 20, width = 16)
ggsave(here("output", DOCNAME, "figure2_panel.pdf"), panel,
       height = 20, width = 16)

panel
```


Figure 3
========

Figure 3A
---------

```{r fig-3A}
plot.data <- comb.neph %>%
    GetDimReduction("tsne",  slot = "cell.embeddings") %>%
    data.frame() %>%
    rownames_to_column("Cell") %>%
    mutate(Group = comb.neph@meta.data$Group) %>%
    mutate(Group = if_else(Group == "Lindstrom", "Fetal kidney", Group))

f3A <- ggplot(plot.data, aes(x = tSNE_1, y = tSNE_2, colour = Group)) +
    geom_point(size = 3) +
    scale_color_manual(values = c("#EC008C", "#00ADEF")) +
    guides(colour = guide_legend(ncol = 4, override.aes = list(size = 12),
                                 label.theme = element_text(size = 12))) +
    theme_cowplot() +
    theme(legend.position = "bottom",
          legend.title = element_blank(),
          legend.justification = "center")

ggsave(here("output", DOCNAME, "figure3A.png"), f3A,
       height = 8, width = 10)
ggsave(here("output", DOCNAME, "figure3A.pdf"), f3A,
       height = 8, width = 10)

f3A
```

Figure 3B
---------

```{r fig-3B}
plot.data <- comb.neph %>%
    GetDimReduction("tsne",  slot = "cell.embeddings") %>%
    data.frame() %>%
    rownames_to_column("Cell") %>%
    mutate(Cluster = comb.neph@ident)

lab.data <- plot.data %>%
    group_by(Cluster) %>%
    summarise(tSNE_1 = mean(tSNE_1),
              tSNE_2 = mean(tSNE_2)) %>%
    mutate(Label = paste0("CN", Cluster))

clust.labs <- c(
    "CN0 (Podocyte)\nTCF21, PODXL, SYNPO",
    "CN1 (Differentiating nephron prog.)\nLYPD1, PAX8, DAPL1",
    "CN2 (Nephron progenitor)\nCRABP2, SIX1, TMEM100, CITED1",
    "CN3 (Podocyte precursor)\nCTGF, OLFM3, MAFB",
    "CN4 (Distal)\nWFDC2, EMX2, MAL, GATA3",
    "CN5 (Proximal)\nIGFBP7, MPC2, FXYD2, CDH6",
    "CN6 (Stroma)\nfetal kidney only",
    "CN7 (Podocyte)\nfetal kidney only"
)

f3B <- ggplot(plot.data, aes(x = tSNE_1, y = tSNE_2, colour = Cluster)) +
    geom_point(size = 3) +
    geom_text(data = lab.data, aes(label = Label), colour = "black", size = 8) +
    #scale_color_brewer(palette = "Set1", labels = clust.labs) +
    scale_colour_discrete(labels = clust.labs) +
    guides(colour = guide_legend(ncol = 4, override.aes = list(size = 12),
                                 label.theme = element_text(size = 12))) +
    theme_cowplot() +
    theme(legend.position = "bottom",
          legend.title = element_blank(),
          legend.justification = "center")

ggsave(here("output", DOCNAME, "figure3B.png"), f3B,
       height = 8, width = 10)
ggsave(here("output", DOCNAME, "figure3B.pdf"), f3B,
       height = 8, width = 10)

f3B
```

Figure 3AB legend
-----------------

```{r fig-3AB-legend}
l3A <- get_legend(f3A)
l3B <- get_legend(f3B)

l3AB <- plot_grid(l3A, l3B, nrow = 1, rel_widths = c(0.3, 1))

l3AB
```

Figure 3C
---------

```{r fig-3C}
f3C <- summariseClusts(clusts, CombNephron, OrgsNephron) %>%
    replace_na(list(Jaccard = 0)) %>%
    ggplot(aes(x = CombNephron, y = OrgsNephron, fill = Jaccard)) +
    geom_tile() +
    scale_fill_viridis_c(limits = c(0, 1), name = "Jaccard\nindex") +
    scale_x_discrete(labels = paste0("CN", 0:7)) +
    scale_y_discrete(labels = paste0("ON", 0:4)) +
    coord_equal() +
    xlab("Combined nephron clusters") +
    ylab("Organoid nephron clusters") +
    theme_minimal() +
    theme(axis.text = element_text(size = 10, colour = "black"),
          axis.ticks = element_blank(),
          axis.title = element_text(size = 16),
          legend.key.height = unit(30, "pt"),
          legend.title = element_text(size = 12),
          legend.text = element_text(size = 10),
          panel.grid = element_blank())

ggsave(here("output", DOCNAME, "figure3C.png"), f3C,
       height = 8, width = 10)
ggsave(here("output", DOCNAME, "figure3C.pdf"), f3C,
       height = 8, width = 10)

f3C
```

Figure 3D
---------

```{r fig-3D}
plot.data <- comb.neph@meta.data %>%
    select(NephCluster, Group) %>%
    mutate(NephCluster = paste0("CN", NephCluster)) %>%
    mutate(Group = if_else(Group == "Lindstrom", "Fetal kidney", Group))

f3D <- ggplot(plot.data, aes(x = NephCluster, fill = NephCluster)) +
    geom_bar() +
    #scale_fill_brewer(palette = "Set1") +
    labs(y = "Number of cells") +
    facet_wrap(~ Group, ncol = 1, scales = "free_y") +
    theme_cowplot() +
    theme(legend.position = "none",
          axis.title.x = element_blank(),
          strip.text = element_text(size = 15,
                                    margin = margin(0, 0, 2, 0, "pt")),
          strip.background = element_blank(),
          strip.placement = "outside")

ggsave(here("output", DOCNAME, "figure3D.png"), f3D,
       height = 8, width = 10)
ggsave(here("output", DOCNAME, "figure3D.pdf"), f3D,
       height = 8, width = 10)

f3D
```

Figure 3E
---------

```{r fig-3E}
genes <- c( "TCF21",   "ROBO2",  "NPHS2", "NPHS1",   "SYNPO",
            "LYPD1",    "PAX8",  "DAPL1",  "CDH6",    "PAX2",   "HEY1",
             "SIX1", "TMEM100", "CITED1", "MEOX1",    "EYA1", "CRABP2",
             "CTGF",   "OLFM3",   "MAFB", "PODXL",   "VAMP8",
            "WFDC2",    "EMX2",    "MAL", "MECOM", "TMEM52B",
           "IGFBP7",   "FXYD2",    "CLU",  "MPC2",
           "COL1A1",  "COL3A1",  "SNAI2", "POSTN",  "LGALS1",
            "PTPRO",   "TPPP3",  "CLIC5", "ENPEP",     "GSN", "TSPAN2")

gene.groups <- rev(c(rep("Podocyte", 5),
                     rep("Differentiating\nnephron progenitor", 6),
                     rep("Nephron\nprogenitor", 6),
                     rep("Podocyte\nprecursor", 5),
                     rep("Distal", 5),
                     rep("Proximal", 4),
                     rep("Stroma", 5),
                     rep("hFK\nPodocyte", 6))) %>%
    fct_relevel("Podocyte", "Differentiating\nnephron progenitor", 
                "Nephron\nprogenitor", "Podocyte\nprecursor", "Distal",
                "Proximal", "Stroma", "hFK\nPodocyte")

clust.labs <- c(
    "hFK CN0", "Org CN0",
    "hFK CN1", "Org CN1",
    "hFK CN2", "Org CN2",
    "hFK CN3", "Org CN3",
    "hFK CN4", "Org CN4",
    "hFK CN5", "Org CN5",
    "hFK CN6", "Org CN6",
    "hFK CN7", "Org CN7"
)

f3E <- SplitDotPlotGG(comb.neph, "Group", genes, gene.groups,
                      cols.use = c("#EC008C", "#00ADEF"), dot.scale = 5,
                      do.return = TRUE) +
    scale_y_discrete(labels = clust.labs) +
    theme(axis.text.x = element_text(angle = 90, vjust = 0.5, hjust = 1),
          strip.text.x = element_text(margin = margin(0, 0, 2, 0, "pt")))

for (y in seq(2.5, 14.5, 2)) {
    f3E <- f3E + geom_hline(yintercept = y, size = 0.2, colour = "grey70")
}

ggsave(here("output", DOCNAME, "figure3E.png"), f3E,
       height = 4, width = 20)
ggsave(here("output", DOCNAME, "figure3E.pdf"), f3E,
       height = 8, width = 10)

f3E
```

Figure 3F
---------

```{r fig-3F}
genes.list <- list(
    c("LYPD1", "PDGFC", "ITGA8", "TUBA1A", "PTPRO", "WT1", "ANXA1", "VAMP8"),
    c("CCND1", "CXCR4", "TUBA1B", "STMN1", "ATOX1", "BNIP3", "GNAS", "FTL"),
    c("TMEM100", "MEOX1", "ROBO2", "DAPL1", "MYOD1", "MYOG", "PITX2", "TPM2"),
    c("HNRNPR", "CXCL1", "OLFM3", "CCND1_1", "PTPRO_1", "WT1_1", "VEGFA",
      "SYNPO"),
    c("PAPPA2", "LIMCH1", "MAL", "PTGR1", "CITED2", "FTL_1", "BNIP3_1",
      "LHX1"),
    c("VCAN", "PSMA2", "CA2", "CCND1_2", "VAMP8_1", "BNIP3_2", "LHX1_1",
      "HNF1B"),
    c("COL1A1", "COL3A1", "SERPINH1", "POSTN", "MEG3", "COL1A2", "CYR61", 
      "MEIS2"),
    c("TNNI1", "TNNT2", "MRGPRF", "MYL9", "COL4A4", "TUBB2A", "HMGCS1", "MSMO1")
)

clust.labs <- c("CN0\nPodocyte", "CN1\nDiff. NP", "CN2\nNP", "CN3\nPod. Pre.",
                "CN4\nDistal", "CN5\nProximal", "CN6\nStroma", "CN7\nPodocyte")

plot.data <- lapply(seq_along(genes.list), function(idx) {
    genes.raw <- genes.list[[idx]]
    genes <- str_remove(genes.raw, "_[0-9]")
    names(genes.raw) <- genes
    comb.neph %>%
        FetchData(vars.all = genes) %>%
        as_data_frame() %>%
        rownames_to_column("Cell") %>%
        mutate(Cluster = as.numeric(as.character(comb.neph@ident)),
               Group = comb.neph@meta.data$Group) %>%
        gather(key = "Gene", value = "Expr", -Cell, -Cluster, -Group) %>%
        group_by(Cluster, Gene, Group) %>%
        summarize(AvgExpr = mean(expm1(Expr)),
                  PctExpr = Seurat:::PercentAbove(Expr, threshold = 0) * 100) %>%
        group_by(Gene) %>%
        mutate(AvgExprScale = scale(AvgExpr)) %>%
        mutate(AvgExprScale = Seurat::MinMax(AvgExprScale,
                                             max = 2.5, min = -2.5)) %>%
        filter(Cluster == idx - 1)  %>%
        ungroup() %>%
        mutate(Gene = genes.raw[Gene])
})

plot.data <- plot.data %>%
    bind_rows() %>%
    mutate(Cluster = factor(Cluster, labels = clust.labs)) %>%
    mutate(Gene = factor(Gene, levels = unlist(genes.list)))

f3F <- ggplot(plot.data,
              aes(x = Group, y = Gene, size = PctExpr,
                  colour = Group, alpha = AvgExprScale)) +
    geom_point() +
    scale_radius(range = c(0, 10)) +
    scale_alpha(range = c(0.1, 1)) +
    scale_colour_manual(values = c("#EC008C", "#00ADEF")) +
    scale_x_discrete(labels = c("Fetal kidney", "Organoid")) +
    scale_y_discrete(labels = str_remove(unlist(genes.list), "_[0-9]"),
                     breaks = unlist(genes.list)) +
    facet_wrap(~ Cluster, scales = "free", nrow = 1) +
    theme(axis.title.x = element_blank(),
          axis.title.y = element_blank(),
          axis.text.x = element_text(angle = 90, vjust = 0.5, hjust = 1),
          panel.spacing = unit(x = 1, units = "lines"),
          strip.background = element_blank(),
          strip.placement = "outside",
          legend.position = "none")

ggsave(here("output", DOCNAME, "figure3F.png"), f3F,
       height = 8, width = 10)
ggsave(here("output", DOCNAME, "figure3F.pdf"), f3F,
       height = 8, width = 10)

f3F
```

Figure 3 Panel
--------------

```{r fig3-panel, fig.height = 20, fig.width = 16}
p1 <- plot_grid(f3A + theme(legend.position = "none"),
                f3B + theme(legend.position = "none"),
                nrow = 1, labels = c("A", "B"),
                label_size = 20)
p2 <- plot_grid(f3C, f3D,
                nrow = 1, labels = c("C", "D"),
                label_size = 20)
p3 <- plot_grid(p1,
                l3AB,
                p2,
                ncol = 1, rel_heights = c(1, 0.2, 0.6))
panel <- plot_grid(p3, f3E, f3F, ncol = 1, labels = c("", "E", "F"),
                   rel_heights = c(1, 0.4, 0.4),
                   label_size = 20)

ggsave(here("output", DOCNAME, "figure3_panel.png"), panel,
       height = 20, width = 16)
ggsave(here("output", DOCNAME, "figure3_panel.pdf"), panel,
       height = 20, width = 16)

panel
```

Summary
=======

Output files
------------

This table describes the output files produced by this document. Right click
and _Save Link As..._ to download the results.

```{r output}
kable(data.frame(
    File = c(
        glue("[figure2A.png]({getDownloadURL('figure2A.png', DOCNAME)})"),
        glue("[figure2A.pdf]({getDownloadURL('figure2A.pdf', DOCNAME)})"),
        glue("[figure2B.png]({getDownloadURL('figure2B.png', DOCNAME)})"),
        glue("[figure2B.pdf]({getDownloadURL('figure2B.pdf', DOCNAME)})"),
        glue("[figure2C.png]({getDownloadURL('figure2C.png', DOCNAME)})"),
        glue("[figure2C.pdf]({getDownloadURL('figure2C.pdf', DOCNAME)})"),
        glue("[figure2D.png]({getDownloadURL('figure2D.png', DOCNAME)})"),
        glue("[figure2D.pdf]({getDownloadURL('figure2D.pdf', DOCNAME)})"),
        glue("[figure2E.png]({getDownloadURL('figure2E.png', DOCNAME)})"),
        glue("[figure2E.pdf]({getDownloadURL('figure2E.pdf', DOCNAME)})"),
        glue("[figure2F.png]({getDownloadURL('figure2F.png', DOCNAME)})"),
        glue("[figure2F.pdf]({getDownloadURL('figure2F.pdf', DOCNAME)})"),
        glue("[figure2G.png]({getDownloadURL('figure2G.png', DOCNAME)})"),
        glue("[figure2G.pdf]({getDownloadURL('figure2G.pdf', DOCNAME)})"),
        glue("[figure2_panel.png]",
             "({getDownloadURL('figure2_panel.png', DOCNAME)})"),
        glue("[figure2_panel.pdf]",
             "({getDownloadURL('figure2_panel.pdf', DOCNAME)})"),
        glue("[figure3A.png]({getDownloadURL('figure3A.png', DOCNAME)})"),
        glue("[figure3A.pdf]({getDownloadURL('figure3A.pdf', DOCNAME)})"),
        glue("[figure3B.png]({getDownloadURL('figure3B.png', DOCNAME)})"),
        glue("[figure3B.pdf]({getDownloadURL('figure3B.pdf', DOCNAME)})"),
        glue("[figure3C.png]({getDownloadURL('figure3C.png', DOCNAME)})"),
        glue("[figure3C.pdf]({getDownloadURL('figure3C.pdf', DOCNAME)})"),
        glue("[figure3D.png]({getDownloadURL('figure3D.png', DOCNAME)})"),
        glue("[figure3D.pdf]({getDownloadURL('figure3D.pdf', DOCNAME)})"),
        glue("[figure3E.png]({getDownloadURL('figure3E.png', DOCNAME)})"),
        glue("[figure3E.pdf]({getDownloadURL('figure3E.pdf', DOCNAME)})"),
        glue("[figure3F.png]({getDownloadURL('figure3F.png', DOCNAME)})"),
        glue("[figure3F.pdf]({getDownloadURL('figure3F.pdf', DOCNAME)})"),
        glue("[figure3_panel.png]",
             "({getDownloadURL('figure3_panel.png', DOCNAME)})"),
        glue("[figure3_panel.pdf]",
             "({getDownloadURL('figure3_panel.pdf', DOCNAME)})")
    ),
    Description = c(
        "Figure 2A in PNG format",
        "Figure 2A in PDF format",
        "Figure 2B in PNG format",
        "Figure 2B in PDF format",
        "Figure 2C in PNG format",
        "Figure 2C in PDF format",
        "Figure 2D in PNG format",
        "Figure 2D in PDF format",
        "Figure 2E in PNG format",
        "Figure 2E in PDF format",
        "Figure 2F in PNG format",
        "Figure 2F in PDF format",
        "Figure 2G in PNG format",
        "Figure 2G in PDF format",
        "Figure 2 panel in PNG format",
        "Figure 2 panel in PDF format",
        "Figure 3A in PNG format",
        "Figure 3A in PDF format",
        "Figure 3B in PNG format",
        "Figure 3B in PDF format",
        "Figure 3C in PNG format",
        "Figure 3C in PDF format",
        "Figure 3D in PNG format",
        "Figure 3D in PDF format",
        "Figure 3E in PNG format",
        "Figure 3E in PDF format",
        "Figure 3F in PNG format",
        "Figure 3F in PDF format",
        "Figure 3 panel in PNG format",
        "Figure 3 panel in PDF format"
    )
))
```

Session information
-------------------

```{r session-info, cache = FALSE}
devtools::session_info()
```
